new healthbar implemented, healthbar prefab made, canvas updated to include health bar, attack/health changed to integer from float

+++ Assets/_Scripts/Entities/Enemies/BossEnemy.cs
@@ -182,7 +182,7 @@ public class BossEnemy : Enemy
     /// <summary>Checks if the entity can be attacked, and attacks them if so.</summary>
     /// <param name="target">The entity to attack.</param>
     /// <param name="damage">The damage to deal to the entity.</param>
-    public override void Attack(Entity target, float damage)
+    public override void Attack(Entity target, int damage)
     {
         base.Attack(target, damage + 5 * Mathf.RoundToInt(bossScaleMult));
     }
